requirements.txt
streamlit
chromadb
numpy
sentence-transformers
google.generativeai
transformers
gTTS
beautifulsoup4
requests
langchain
